avbus: setSource: and currentID return unsigned short, not unsigned int
Both were declared unsigned int. Reading their method_t types strings out of
the binary gives S24@0:8^{AVSource=@@B}16 and S16@0:8, so both return S, an
unsigned short, and both bodies return the S-encoded mCurrentID through a
16-bit ldrh. The width rule asks for the encoding's own spelling.

The two callers in avsemixer need no change. One compares against a masked
handle and the other composes dwCurrentId into a 32-bit handle, so the
narrowed return promotes cleanly in both.

-setSource: itself is verified: it stores the record, dispatches on the first
field with the URL winning when set and the data buffer otherwise, passes the
+0x10 BOOL through as isLoop, and returns mCurrentID. RBMenuView
-SelectPlaylistFinButton synchronises the playlist manager and tail-calls
playlistEditFinish.

+++ Project/GameSystem/src/Sound/AVBus.h
@@ -45,7 +45,7 @@ typedef NS_ENUM(int, AVBusStatus) {
  * @return The voice's current id, packed into the low half of a play handle by the mixer.
  * @ghidraAddress 0x4169c
  */
-- (unsigned int)setSource:(AudioSourceSlot::SourceRecord *)source;
+- (unsigned short)setSource:(AudioSourceSlot::SourceRecord *)source;
 /**
  * @brief Unbinds the voice's current source and invalidates its outstanding play handles.
  * @return @c YES when a player was released.
@@ -101,7 +101,7 @@ typedef NS_ENUM(int, AVBusStatus) {
  * @brief The voice's current id, matched against a play handle's low half to resolve the voice.
  * @ghidraAddress 0x41f38
  */
-- (unsigned int)currentID;
+- (unsigned short)currentID;
 /**
  * @brief The voice's playback state.
  * @ghidraAddress 0x41d04
